+++ src/ontodia/data/sparql/sparqlDataProvider.ts
@@ -15,10 +15,11 @@ import {
     getLinkTypesInfo,
 } from './responseHandler';
 import {
-    ClassBinding, ElementBinding, LinkBinding, PropertyBinding,
-    LinkCountBinding, LinkTypeBinding, ElementImageBinding, SparqlResponse, Triple, RdfNode,
+    ClassBinding, ElementBinding, LinkBinding, PropertyBinding, BlankBinding, isBlankBinding,
+    LinkCountBinding, LinkTypeBinding, ElementImageBinding, SparqlResponse, Triple, RdfNode, isRdfBlank,
 } from './sparqlModels';
 import { SparqlDataProviderSettings, OWLStatsSettings } from './sparqlDataProviderSettings';
+import * as BlankNodes from './blankNodes';
 
 export enum SparqlQueryMethod { GET = 1, POST }
 
@@ -33,6 +34,8 @@ export type QueryFunction = (params: {
  * Runtime settings of SPARQL data provider
  */
 export interface SparqlDataProviderOptions {
+
+    acceptBlankNodes?: boolean;
 
     /**
      *  sparql endpoint URL to use
@@ -59,26 +62,22 @@ export interface SparqlDataProviderOptions {
     queryMethod?: SparqlQueryMethod;
 
     /*
-     * what property to use as instance labels. This will override dataLabelProperty from settings
-     */
-    labelProperty?: string;
-
-    /*
      * function to send sparql requests
      */
     queryFunction?: QueryFunction;
 }
 
 export class SparqlDataProvider implements DataProvider {
-    dataLabelProperty: string;
+    readonly options: SparqlDataProviderOptions;
+    readonly settings: SparqlDataProviderSettings;
+
     constructor(
-        private options: SparqlDataProviderOptions,
-        private settings: SparqlDataProviderSettings = OWLStatsSettings
+        options: SparqlDataProviderOptions,
+        settings: SparqlDataProviderSettings = OWLStatsSettings,
     ) {
-        this.dataLabelProperty = options.labelProperty ? options.labelProperty : settings.dataLabelProperty;
-        if (this.options.queryFunction == null) {
-            this.options.queryFunction = queryInternal;
-        }
+        const {queryFunction = queryInternal} = options;
+        this.options = {...options, queryFunction};
+        this.settings = settings;
     }
 
     classTree(): Promise<ClassModel[]> {
@@ -136,10 +135,24 @@ export class SparqlDataProvider implements DataProvider {
     }
 
     elementInfo(params: { elementIds: string[]; }): Promise<Dictionary<ElementModel>> {
-        const ids = params.elementIds.map(escapeIri).map(id => ` (${id})`).join(' ');
-        const query = this.settings.defaultPrefix
-            + resolveTemplate(this.settings.elementInfoQuery, {ids: ids, dataLabelProperty: this.dataLabelProperty});
-        return this.executeSparqlQuery<ElementBinding>(query)
+        const blankIds: string[] = [];
+
+        const elementIds = params.elementIds.filter(id => !BlankNodes.isBlunkNodeId(id));
+        const additionalData: SparqlResponse<ElementBinding> = (
+            this.options.acceptBlankNodes ?
+            BlankNodes.elementInfo(params.elementIds) :
+            null
+        );
+
+        if (elementIds.length === 0 && this.options.acceptBlankNodes) {
+            return Promise.resolve(getElementsInfo(additionalData, params.elementIds));
+        }
+
+        const ids = elementIds.map(escapeIri).map(id => ` (${id})`).join(' ');
+        const {defaultPrefix, dataLabelProperty, elementInfoQuery} = this.settings;
+        const query = defaultPrefix + resolveTemplate(elementInfoQuery, {ids, dataLabelProperty});
+
+        return this.executeSparqlQuery<ElementBinding>(query, {additionalData})
             .then(elementsInfo => getElementsInfo(elementsInfo, params.elementIds))
             .then(elementModels => {
                 if (this.options.prepareImages) {
@@ -154,9 +167,10 @@ export class SparqlDataProvider implements DataProvider {
 
     private enrichedElementsInfo(
         elementsInfo: Dictionary<ElementModel>,
-        types: string[]
+        types: string[],
     ): Promise<Dictionary<ElementModel>> {
-        const ids = Object.keys(elementsInfo).map(escapeIri).map(id => ` ( ${id} )`).join(' ');
+        const ids = Object.keys(elementsInfo).filter(id => !BlankNodes.isBlunkNodeId(id))
+            .map(escapeIri).map(id => ` ( ${id} )`).join(' ');
         const typesString = types.map(escapeIri).map(id => ` ( ${id} )`).join(' ');
 
         const query = this.settings.defaultPrefix + `
@@ -176,7 +190,7 @@ export class SparqlDataProvider implements DataProvider {
     }
 
     private prepareElementsImage(
-        elementsInfo: Dictionary<ElementModel>
+        elementsInfo: Dictionary<ElementModel>,
     ): Promise<Dictionary<ElementModel>> {
         return this.options.prepareImages(elementsInfo).then(images => {
             for (const key in images) {
@@ -192,17 +206,37 @@ export class SparqlDataProvider implements DataProvider {
         elementIds: string[];
         linkTypeIds: string[];
     }): Promise<LinkModel[]> {
-        const ids = params.elementIds.map(escapeIri).map(id => ` ( ${id} )`).join(' ');
+        const elementIds = params.elementIds.filter(id => !BlankNodes.isBlunkNodeId(id));
+
+        const additionalData = (
+            this.options.acceptBlankNodes ?
+            BlankNodes.linksInfo(params.elementIds) :
+            null
+        );
+
+        if (elementIds.length === 0 && this.options.acceptBlankNodes) {
+            return Promise.resolve(getLinksInfo(additionalData));
+        }
+
+        const ids = elementIds.map(escapeIri).map(id => ` ( ${id} )`).join(' ');
         const linksInfoQuery = resolveTemplate(this.settings.linksInfoQuery, {ids: ids});
         const query = this.settings.defaultPrefix + linksInfoQuery;
-        return this.executeSparqlQuery<LinkBinding>(query).then(getLinksInfo);
+        return this.executeSparqlQuery<LinkBinding>(query, {additionalData}).then(getLinksInfo);
     }
 
     linkTypesOf(params: { elementId: string; }): Promise<LinkCount[]> {
-        const elementIri = escapeIri(params.elementId);
-        const query = this.settings.defaultPrefix
-            + resolveTemplate(this.settings.linkTypesOfQuery, {elementIri: elementIri});
-        return this.executeSparqlQuery<LinkCountBinding>(query).then(getLinksTypesOf);
+        if (this.options.acceptBlankNodes && BlankNodes.isBlunkNodeId(params.elementId)) {
+            return Promise.resolve(getLinksTypesOf(BlankNodes.linkTypesOf(params)));
+        } else {
+            const elementIri = escapeIri(params.elementId);
+            const query = this.settings.defaultPrefix
+                + resolveTemplate(this.settings.linkTypesOfQuery, {elementIri: elementIri});
+            return this.executeSparqlQuery<LinkCountBinding>(query).
+                then(result => {
+                    result.results.bindings = result.results.bindings.filter(b => !isRdfBlank(b.link));
+                    return result;
+                }).then(getLinksTypesOf);
+        }
     };
 
     linkElements(params: {
@@ -224,105 +258,147 @@ export class SparqlDataProvider implements DataProvider {
 
     filter(params: FilterParams): Promise<Dictionary<ElementModel>> {
         if (params.limit === 0) { params.limit = 100; }
+        const filterData = (
+            this.options.acceptBlankNodes ?
+            BlankNodes.filter(params) :
+            null
+        );
+
+        if (this.options.acceptBlankNodes && Object.keys(filterData.results.bindings).length > 0) {
+            return Promise.resolve(getFilteredData(filterData));
+        }
 
         if (!params.refElementId && params.refElementLinkId) {
             throw new Error(`Can't execute refElementLink filter without refElement`);
         }
 
-        let refQueryPart = this.createRefQueryPart({
-            elementId: params.refElementId,
-            linkId: params.refElementLinkId,
-            direction: params.linkDirection,
-        });
+        let refQueryPart = '';
+        if (params.refElementId) {
+            refQueryPart = this.createRefQueryPart({
+                elementId: params.refElementId,
+                linkId: params.refElementLinkId,
+                direction: params.linkDirection,
+            });
+        }
 
-        let elementTypePart: string;
+        let elementTypePart = '';
         if (params.elementTypeId) {
             const elementTypeIri = escapeIri(params.elementTypeId);
-            elementTypePart = resolveTemplate(this.settings.filterTypePattern, {elementTypeIri: elementTypeIri});
-        } else {
-            elementTypePart = '';
+            elementTypePart = resolveTemplate(this.settings.filterTypePattern, {elementTypeIri});
         }
 
-        let textSearchPart: string;
+        let textSearchPart = '';
         if (params.text) {
-            const text = params.text;
-            textSearchPart = resolveTemplate(
-                this.settings.fullTextSearch.queryPattern,
-                {text: text, dataLabelProperty: this.dataLabelProperty}
-            );
-        } else {
-            textSearchPart = '';
+            const {fullTextSearch, dataLabelProperty} = this.settings;
+            textSearchPart = resolveTemplate(fullTextSearch.queryPattern, {text: params.text, dataLabelProperty});
         }
 
-        let query = `${this.settings.defaultPrefix}
-            ${this.settings.fullTextSearch.prefix}
+        const {defaultPrefix, fullTextSearch, dataLabelProperty} = this.settings;
+        const query = `${defaultPrefix}
+            ${fullTextSearch.prefix}
             
-        SELECT ?inst ?class ?label
+        SELECT ?inst ?class ?label ?blankTrgProp ?blankTrg ?blankSrc ?blankSrcProp ?blankType ?listHead
         WHERE {
             {
-                SELECT DISTINCT ?inst ?score WHERE {
+                SELECT DISTINCT ?inst ${textSearchPart ? '?score' : ''} WHERE {
                     ${elementTypePart}
                     ${refQueryPart}
                     ${textSearchPart}
                     ${this.settings.filterAdditionalRestriction}
                     ${this.settings.fullTextSearch.extractLabel ? sparqlExtractLabel('?inst', '?extractedLabel') : ''}
-                } ORDER BY DESC(?score) LIMIT ${params.limit} OFFSET ${params.offset}
+                }
+                ${textSearchPart ? 'ORDER BY DESC(?score)' : ''}
+                LIMIT ${params.limit} OFFSET ${params.offset}
             }
-            ${resolveTemplate(this.settings.filterElementInfoPattern, {dataLabelProperty: this.dataLabelProperty})}
-        } ORDER BY DESC(?score)
+            ${resolveTemplate(this.settings.filterElementInfoPattern, {dataLabelProperty})}
+            OPTIONAL {
+                FILTER (ISBLANK(?inst)).
+                {
+                    ?inst ?blankTrgProp ?blankTrg.
+                    ?blankSrc ?blankSrcProp ?inst.
+                    FILTER NOT EXISTS { ?inst rdf:first _:smth1 }.
+                    BIND("blankNode" as ?blankType)
+                } UNION {
+                    ?inst rdf:rest*/rdf:first ?blankTrg.
+                    ?blankSrc ?blankSrcProp ?inst.
+                    _:smth2 rdf:first ?blankTrg.
+                    BIND(?blankSrcProp as ?blankTrgProp)
+                    BIND("listHead" as ?blankType)
+                    FILTER NOT EXISTS { _:smth3 rdf:rest ?inst }.
+                } UNION {
+                    ?listHead rdf:rest* ?inst.
+                    FILTER NOT EXISTS { _:smth4 rdf:rest ?listHead }.
+
+                    ?listHead rdf:rest*/rdf:first ?blankTrg.
+                    ?blankSrc ?blankSrcProp ?listHead.
+                    _:smth5 rdf:first ?blankTrg.
+                    BIND(?blankSrcProp as ?blankTrgProp)
+                    BIND("listHead" as ?blankType)
+                }
+            }
+        } ${textSearchPart ? 'ORDER BY DESC(?score)' : ''}
         `;
 
-        return this.executeSparqlQuery<ElementBinding>(query).then(getFilteredData);
+        return this.executeSparqlQuery<ElementBinding | BlankBinding>(query, {
+            additionalData: filterData,
+        }).then(result => {
+            if (this.options.acceptBlankNodes) {
+                return BlankNodes.updateFilterResults(
+                    result,
+                    (callBackQuery: string) => {
+                        return this.executeSparqlQuery<BlankBinding>(callBackQuery);
+                    },
+                );
+            } else {
+                result.results.bindings = result.results.bindings.filter(b => !isBlankBinding(b));
+                return result;
+            }
+        }).then(getFilteredData);
     };
 
-    executeSparqlQuery<Binding>(query: string) {
+    executeSparqlQuery<Binding>(query: string, params?: { additionalData: SparqlResponse<Binding> }) {
         const method = this.options.queryMethod ? this.options.queryMethod : SparqlQueryMethod.GET;
-        return executeSparqlQuery<Binding>(this.options.endpointUrl, query, method, this.options.queryFunction);
+        let promise = executeSparqlQuery<Binding>(this.options.endpointUrl, query, method, this.options.queryFunction);
+        if (params && params.additionalData) {
+            promise = promise.then(result => {
+                result.results.bindings = result.results.bindings.concat(params.additionalData.results.bindings);
+                return result;
+            });
+        }
+        return promise;
     }
 
-    executeSparqlConstruct(query: string) : Promise<Triple[]> {
+    executeSparqlConstruct(query: string): Promise<Triple[]> {
         const method = this.options.queryMethod ? this.options.queryMethod : SparqlQueryMethod.GET;
         return executeSparqlConstruct(this.options.endpointUrl, query, method, this.options.queryFunction);
     }
 
-    createRefQueryPart(params: { elementId: string; linkId?: string; direction?: 'in' | 'out'}) {
-        let refQueryPart = '';
+    protected createRefQueryPart(params: { elementId: string; linkId?: string; direction?: 'in' | 'out'}) {
+        const {elementId, linkId, direction} = params;
         const refElementIRI = escapeIri(params.elementId);
-        const refElementLinkIRI = params.linkId ? escapeIri(params.linkId) : undefined;
-
+        const linkPattern = linkId ? escapeIri(params.linkId) : '?link';
+        const blankFilter = this.options.acceptBlankNodes
+            ? 'FILTER(isIri(?inst) || isBlank(?inst))'
+            : 'FILTER(isIri(?inst))';
         // link to element with specified link type
         // if direction is not specified, provide both patterns and union them
         // FILTER ISIRI is used to prevent blank nodes appearing in results
-        if (params.elementId && params.linkId) {
-            const refElementLinkQueryOut = resolveTemplate(
-                this.settings.refElementLinkQueryOut,
-                {refElementIRI: refElementIRI, refElementLinkIRI: refElementLinkIRI},
-            );
-            const refElementLinkQueryIn = resolveTemplate(
-                this.settings.refElementLinkQueryIn,
-                {refElementIRI: refElementIRI, refElementLinkIRI: refElementLinkIRI},
-            );
-            refQueryPart += !params.direction || params.direction === 'out' ? `{ ${refElementLinkQueryOut} }` : '';
-            refQueryPart += !params.direction ? ' UNION ' : '';
-            refQueryPart += !params.direction || params.direction === 'in' ? `{ ${refElementLinkQueryIn} }` : '';
+        let part = '';
+        if (params.direction !== 'in') {
+            part += `{ ${refElementIRI} ${linkPattern} ?inst . ${blankFilter} }`;
         }
-
-        // all links to current element
-        if (params.elementId && !params.linkId) {
-            const refElementQueryOut = resolveTemplate(this.settings.refElementQueryOut, {refElementIRI: refElementIRI});
-            const refElementQueryIn = resolveTemplate(this.settings.refElementQueryIn, {refElementIRI: refElementIRI});
-
-            refQueryPart += !params.direction || params.direction === 'out' ? `{ ${refElementQueryOut} }` : '';
-            refQueryPart += !params.direction ? ' UNION ' : '';
-            refQueryPart += !params.direction || params.direction === 'in' ? `{ ${refElementQueryIn} }` : '';
+        if (!params.direction) { part += ' UNION '; }
+        if (params.direction !== 'out') {
+            part += `{ ?inst ${linkPattern} ${refElementIRI} . ${blankFilter} }`;
         }
-        return refQueryPart;
+        return part;
     }
 }
 
 function resolveTemplate(template: string, values: Dictionary<string>) {
     let result = template;
     for (const replaceKey in values) {
+        if (!values.hasOwnProperty(replaceKey)) { continue; }
         const replaceValue = values[replaceKey];
         result = result.replace(new RegExp('\\${' + replaceKey + '}', 'g'), replaceValue);
     }
@@ -464,6 +540,9 @@ function sparqlExtractLabel(subject: string, label: string): string {
 }
 
 function escapeIri(iri: string) {
+    if (typeof iri !== 'string') {
+        throw new Error(`Cannot escape IRI of type "${typeof iri}"`);
+    }
     return `<${iri}>`;
 }